Add a script that renames all photo's and movies in a dir to date/time.
When photo's and movies from multiple sources are thrown in one dir it
often becomes a mess of different file formats. Also it's impossible to
view the images in chronological order. Very annoying when it's hundreds
of holiday photo's of both my wife and me.
This script searches for information about the date/time the photo/movie
was taken/shot and renames the file to `date_time.ext`.